Add correct api form and also averagePopularity in logic

+++ CA_05_mzFoodDelivery/server/CA_05_mzFoodDelivery/src/main/java/ir/ac/ut/ie/CA_05_mzFoodDelivery/domain/MzFoodDelivery/Restaurant/Restaurant.java
@@ -84,8 +84,8 @@ public class Restaurant {
         menu.add(food);
     }
 
-    public int getFoodsPopularityAverage() {
-        int foodsPopulationSum = 0;
+    public double getFoodsPopularityAverage() {
+        double foodsPopulationSum = 0;
         for (Food food: menu) {
             foodsPopulationSum += food.getPopularity();
         }
